Add note about side effects to Filter docs
Issue #30632 shows the unexpected behavior that can arise from
`iter::Map` and `iter::Filter` both implementing `DoubleEndedIterator`.
PR #31220 makes note of this behavior by adding a note to the docs for
`iter::Map`, but does not do the same for `iter::Filter`. This PR adds
similar documentation to `iter::Filter`.

It may be worth considering adding this documentation to
`iter::FilterMap` as well or placing it in another location with a note
that it applies to using all of these types with stateful closures.